requirements.txt
Mako==1.3.5	
MarkupSafe==2.1.5	
PyYAML==6.0.1	
SQLAlchemy==2.0.30	
alembic==1.13.1	
annotated-types==0.7.0
anyio==4.4.0	
bcrypt==4.1.3
certifi==2024.6.2
click==8.1.7	
colorama==0.4.6	
dnspython==2.6.1	
ecdsa==0.19.0	
email-validator==2.1.1	
exceptiongroup==1.2.1	
fastapi==0.111.0	
fastapi-cli==0.0.4	
greenlet==3.0.3	
h11==0.14.0	
httpcore==1.0.5	
httptools==0.6.1	
httpx==0.27.0	
idna==3.7
jinja2==3.1.4	
markdown-it-py==3.0.0	
mdurl==0.1.2
orjson==3.10.3	
passlib==1.7.4	
pip==23.2.1	
psycopg2==2.9.9	
pyasn1==0.6.0	
pydantic==2.7.3	
pydantic-core==2.18.4	
pydantic-settings==2.3.3	
pygments==2.18.0	
python-dotenv==1.0.1	
python-jose==3.3.0	
python-multipart==0.0.9	
rich==13.7.1	
rsa==4.9	
setuptools==68.2.0	
shellingham==1.5.4
six==1.16.0	
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3	
typing-extensions==4.12.1
ujson==5.10.0	
uvicorn==0.30.1	
watchfiles==0.22.0	
websockets==12.0	
wheel==0.41.2	